+++ waspc/data/packages/spec/src/spec-pipeline/transform/specPackageImports.ts
@@ -1,4 +1,5 @@
 import * as ts from "typescript";
+import { getImportStatements } from "./importStatements.js";
 
 export const SPEC_PACKAGE_NAME = "@wasp.sh/spec";
 export const SPEC_PACKAGE_INTERNAL_NAME = "@wasp.sh/spec/internal";
@@ -8,9 +9,7 @@ export const REF_IMPORT_FACTORY_EXPORT_NAME = "_waspMakeRef";
 export function getSpecPackageImports(
   sourceFile: ts.SourceFile,
 ): ts.ImportDeclaration[] {
-  return sourceFile.statements
-    .filter(ts.isImportDeclaration)
-    .filter(isSpecPackageImport);
+  return getImportStatements(sourceFile).filter(isSpecPackageImport);
 }
 
 function isSpecPackageImport(stmt: ts.ImportDeclaration): boolean {
@@ -24,60 +23,9 @@ function isSpecPackageSpecifier(moduleSpecifier: string): boolean {
   return moduleSpecifier === SPEC_PACKAGE_NAME;
 }
 
-export function getLocalNameForValueImport(
-  imports: ts.ImportDeclaration[],
-  exportedName: string,
-): string | undefined {
-  for (const stmt of imports) {
-    const specifier = getNamedValueImports(stmt)?.elements.find((specifier) =>
-      isValueImportSpecifierFor(specifier, exportedName),
-    );
-
-    if (specifier !== undefined) {
-      return specifier.name.text;
-    }
-  }
-
-  return undefined;
-}
-
-export function getNamedImports(
-  stmt: ts.ImportDeclaration,
-): ts.NamedImports | undefined {
-  const namedBindings = stmt.importClause?.namedBindings;
-  return namedBindings && ts.isNamedImports(namedBindings)
-    ? namedBindings
-    : undefined;
-}
-
-export function getNamedValueImports(
-  stmt: ts.ImportDeclaration,
-): ts.NamedImports | undefined {
-  const importClause = stmt.importClause;
-  if (!importClause || importClause.isTypeOnly) {
-    return undefined;
-  }
-
-  return getNamedImports(stmt);
-}
-
-export function isValueImportSpecifierFor(
-  specifier: ts.ImportSpecifier,
-  exportedName: string,
-): boolean {
-  return !specifier.isTypeOnly && isImportSpecifierFor(specifier, exportedName);
-}
-
-export function isImportSpecifierFor(
-  specifier: ts.ImportSpecifier,
-  exportedName: string,
-): boolean {
-  return getImportedName(specifier) === exportedName;
-}
-
 type ImportSpecifierSource = ts.ImportSpecifier | string;
 
-export function getSpecPackageImportSource(
+export function getSpecPackageImportStatementSource(
   sourceFile: ts.SourceFile,
   stmt: ts.ImportDeclaration,
   namedSpecifiers: ImportSpecifierSource[],
@@ -102,8 +50,4 @@ export function getSpecPackageImportSource(
   ].filter((part): part is string => part !== undefined);
 
   return `import ${importParts.join(", ")} from ${stmt.moduleSpecifier.getText(sourceFile)};`;
-}
-
-function getImportedName(specifier: ts.ImportSpecifier): string {
-  return specifier.propertyName?.text ?? specifier.name.text;
 }
